backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.10.0
pydantic-settings
python-dotenv
httpx
ollama
python-multipart